+++ mqmetric/globals.go
@@ -1,7 +1,7 @@
 package mqmetric
 
 /*
-  Copyright (c) IBM Corporation 2016, 2022
+  Copyright (c) IBM Corporation 2016, 2023
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -50,6 +50,7 @@ type connectionInfo struct {
 	usePublications      bool
 	useStatus            bool
 	useResetQStats       bool
+	useDepthFromStatus   bool
 	showInactiveChannels bool
 	hideSvrConnJobname   bool
 	hideAMQPClientId     bool